requirements.txt
python-telegram-bot>=21.0
psycopg2-binary>=2.9.9
openai>=1.51.0


